feat: make eeg viewer a live stream & download dataset after recording
